🚚 rename: destination 데이터 분리
